# docs/api.yaml
openapi: 3.0.0
info:
  title: 智能个人财务记账系统 API
  version: 2.0.0
  description: |
    一个支持微信账单导入、智能分类、预算管理的个人财务记账系统。

    ## 认证方式
    除注册、登录接口外，所有接口需在请求头中携带 Bearer Token：
    ```
    Authorization: Bearer <access_token>
    ```

    ## 响应格式
    所有接口返回统一的 JSON 格式：
    ```json
    {
      "code": 200,
      "message": "success",
      "data": {}
    }
    ```

servers:
  - url: http://localhost:8000
    description: 开发服务器

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证

  schemas:
    # 通用响应
    Response:
      type: object
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: success
        data:
          oneOf:
            - type: object
            - type: array
            - type: null
          description: 业务数据

    # 用户相关
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: 用户名
        email:
          type: string
          format: email
          description: 邮箱
        password:
          type: string
          minLength: 6
          maxLength: 100
          description: 密码

    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名或邮箱
        password:
          type: string
          description: 密码

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
        username:
          type: string
          description: 用户名
        email:
          type: string
          description: 邮箱
        avatar:
          type: string
          nullable: true
          description: 头像URL
        is_active:
          type: boolean
          description: 是否激活
        created_at:
          type: string
          format: date-time
          description: 创建时间

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: 访问令牌
        token_type:
          type: string
          default: bearer
          description: 令牌类型
        expires_in:
          type: integer
          description: 过期时间（秒）

    # 分类相关
    CategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: 分类名称
        type:
          type: string
          enum: [income, expense]
          default: expense
          description: 分类类型
        icon:
          type: string
          maxLength: 10
          nullable: true
          description: 图标
        color:
          type: string
          maxLength: 7
          nullable: true
          description: 颜色
        parent_id:
          type: integer
          nullable: true
          description: 父分类ID
        sort_order:
          type: integer
          minimum: 0
          default: 0
          description: 排序序号

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        type:
          type: string
          enum: [income, expense]
        icon:
          type: string
          maxLength: 10
        color:
          type: string
          maxLength: 7
        parent_id:
          type: integer
          nullable: true
        sort_order:
          type: integer
          minimum: 0

    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          description: 分类ID
        name:
          type: string
          description: 分类名称
        type:
          type: string
          enum: [income, expense]
          description: 分类类型
        icon:
          type: string
          nullable: true
          description: 图标
        color:
          type: string
          nullable: true
          description: 颜色
        parent_id:
          type: integer
          nullable: true
          description: 父分类ID
        sort_order:
          type: integer
          description: 排序序号
        is_system:
          type: boolean
          description: 是否系统分类
        created_at:
          type: string
          format: date-time
          nullable: true
          description: 创建时间

    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        total:
          type: integer
          description: 总数

    # 账户相关
    AccountCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: 账户名称
        type:
          type: string
          enum: [cash, bank, wechat, alipay, meal_card, other]
          default: other
          description: 账户类型
        initial_balance:
          type: number
          format: decimal
          default: 0
          description: 初始余额
        icon:
          type: string
          maxLength: 10
          nullable: true
          description: 图标
        color:
          type: string
          maxLength: 7
          nullable: true
          description: 颜色
        is_default:
          type: boolean
          default: false
          description: 是否默认账户
        description:
          type: string
          nullable: true
          description: 账户描述

    AccountUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        type:
          type: string
          enum: [cash, bank, wechat, alipay, meal_card, other]
        icon:
          type: string
          maxLength: 10
        color:
          type: string
          maxLength: 7
        is_default:
          type: boolean
        is_enabled:
          type: boolean
        description:
          type: string
          nullable: true

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          description: 账户ID
        name:
          type: string
          description: 账户名称
        type:
          type: string
          enum: [cash, bank, wechat, alipay, meal_card, other]
          description: 账户类型
        balance:
          type: number
          format: decimal
          description: 当前余额
        initial_balance:
          type: number
          format: decimal
          description: 初始余额
        icon:
          type: string
          nullable: true
          description: 图标
        color:
          type: string
          nullable: true
          description: 颜色
        is_default:
          type: boolean
          description: 是否默认账户
        is_enabled:
          type: boolean
          description: 是否启用
        description:
          type: string
          nullable: true
          description: 账户描述
        created_at:
          type: string
          format: date-time
          nullable: true
          description: 创建时间

    AccountListResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        total:
          type: integer
          description: 总数

    AccountSummaryResponse:
      type: object
      properties:
        total_balance:
          type: number
          format: decimal
          description: 总资产
        total_accounts:
          type: integer
          description: 账户总数
        account_distribution:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                description: 账户类型
              balance:
                type: number
                description: 余额
              count:
                type: integer
                description: 账户数量

    TransferRequest:
      type: object
      required:
        - from_account_id
        - to_account_id
        - amount
      properties:
        from_account_id:
          type: integer
          description: 转出账户ID
        to_account_id:
          type: integer
          description: 转入账户ID
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
          description: 转账金额
        remark:
          type: string
          nullable: true
          description: 转账备注
        transaction_date:
          type: string
          format: date-time
          nullable: true
          description: 转账时间

    BalanceAdjustRequest:
      type: object
      required:
        - new_balance
      properties:
        new_balance:
          type: number
          format: decimal
          description: 调整后余额
        remark:
          type: string
          nullable: true
          description: 调整原因

paths:
  # ==================== 认证模块 ====================
  /api/auth/register:
    post:
      summary: 用户注册
      tags: [认证]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          description: 用户名或邮箱已存在

  /api/auth/login:
    post:
      summary: 用户登录
      tags: [认证]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 用户名或密码错误

  /api/auth/me:
    get:
      summary: 获取当前用户信息
      tags: [认证]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'

  # ==================== 分类管理 ====================
  /api/categories:
    get:
      summary: 获取分类列表
      tags: [分类管理]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
          description: 分类类型
        - name: include_system
          in: query
          schema:
            type: boolean
            default: true
          description: 是否包含系统分类
        - name: parent_id
          in: query
          schema:
            type: integer
          description: 父分类ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CategoryListResponse'

    post:
      summary: 创建分类
      tags: [分类管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 分类名称已存在

  /api/categories/tree:
    get:
      summary: 获取分类树
      tags: [分类管理]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
          description: 分类类型
        - name: include_system
          in: query
          schema:
            type: boolean
            default: true
          description: 是否包含系统分类
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /api/categories/stats:
    get:
      summary: 获取带统计的分类列表
      tags: [分类管理]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
          description: 分类类型
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 返回数量限制
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /api/categories/{category_id}:
    get:
      summary: 获取分类详情
      tags: [分类管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: 分类不存在

    put:
      summary: 更新分类
      tags: [分类管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: 更新成功
        '400':
          description: 系统分类不可修改
        '404':
          description: 分类不存在

    delete:
      summary: 删除分类
      tags: [分类管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '400':
          description: 系统分类不可删除/有子分类/有交易记录
        '404':
          description: 分类不存在

  # ==================== 账户管理 ====================
  /api/accounts:
    get:
      summary: 获取账户列表
      tags: [账户管理]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [cash, bank, wechat, alipay, meal_card, other]
          description: 账户类型
        - name: is_enabled
          in: query
          schema:
            type: boolean
          description: 是否启用
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountListResponse'

    post:
      summary: 创建账户
      tags: [账户管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountResponse'
        '400':
          description: 账户名称已存在

  /api/accounts/summary:
    get:
      summary: 获取账户统计摘要
      tags: [账户管理]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountSummaryResponse'

  /api/accounts/default:
    get:
      summary: 获取默认账户
      tags: [账户管理]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountResponse'
        '404':
          description: 没有可用账户

  /api/accounts/transfer:
    post:
      summary: 账户间转账
      tags: [账户管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: 转账成功
        '400':
          description: 余额不足/同一账户
        '404':
          description: 账户不存在

  /api/accounts/{account_id}:
    get:
      summary: 获取账户详情
      tags: [账户管理]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountResponse'
        '404':
          description: 账户不存在

    put:
      summary: 更新账户
      tags: [账户管理]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: 更新成功
        '404':
          description: 账户不存在

    delete:
      summary: 删除账户
      tags: [账户管理]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '400':
          description: 有交易记录无法删除
        '404':
          description: 账户不存在

  /api/accounts/{account_id}/adjust-balance:
    post:
      summary: 调整账户余额
      tags: [账户管理]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BalanceAdjustRequest'
      responses:
        '200':
          description: 调整成功
        '404':
          description: 账户不存在